# openapi.yaml
openapi: 3.1.1
info:
  title: Blogging Platform API
  description: A multi-user blogging platform built with Next.js, tRPC, and PostgreSQL
  version: 1.0.0
  contact:
    name: API Support
    email: support@blogplatform.com
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-production-url.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/cleanup/expired-registrations:
    post:
      summary: Cleanup expired pending registrations
      description: Removes all expired OTP registration requests from the database. This endpoint is intended for automated cleanup jobs.
      tags: [Maintenance]
      security:
        - bearerAuth: []  # Optional API key authentication
      responses:
        '200':
          description: Cleanup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Expired pending registrations cleaned up successfully"
                  timestamp:
                    type: string
                    format: date-time
        '401':
          description: Unauthorized - Invalid or missing API key
        '500':
          description: Internal server error during cleanup
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...